moving widget to correct place
