+++ src/languages/en.ts
@@ -5069,7 +5069,7 @@ const translations = {
         },
         moneyRequestReport: {
             emptyStateTitle: 'This report has no expenses',
-            emptyStateSubtitle: 'You can add expenses to this repor \n using the button above.',
+            emptyStateSubtitle: 'You can add expenses to this report \n using the button above.',
         },
         noCategory: 'No category',
         noTag: 'No tag',
